fix(striker-ui-api): break apart shell
