+++ striker-ui-api/src/lib/shell/index.ts
@@ -1,0 +1,15 @@
+export * from './date';
+export * from './getent';
+export * from './mkfifo';
+export * from './openssl';
+export * from './perr';
+export * from './pout';
+export * from './poutvar';
+export * from './print';
+export * from './resolveGid';
+export * from './resolveId';
+export * from './resolveUid';
+export * from './rm';
+export * from './systemCall';
+export * from './uuid';
+export * from './uuidgen';
