+++ striker-ui-api/src/lib/shell/mkfifo.ts
@@ -1,0 +1,6 @@
+import { SERVER_PATHS } from '../consts';
+
+import { systemCall } from './systemCall';
+
+export const mkfifo = (...args: string[]) =>
+  systemCall(SERVER_PATHS.usr.bin.mkfifo.self, args);
