+++ striker-ui-api/src/lib/shell/openssl.ts
@@ -1,0 +1,6 @@
+import { SERVER_PATHS } from '../consts';
+
+import { systemCall } from './systemCall';
+
+export const openssl = (...args: string[]) =>
+  systemCall(SERVER_PATHS.usr.bin.openssl.self, args);
