+++ striker-ui-api/src/lib/shell/pout.ts
@@ -1,0 +1,7 @@
+import { DEBUG_MAIN } from '../consts';
+
+import { print } from './print';
+
+export const pout = DEBUG_MAIN
+  ? (message: string) => print(message)
+  : () => null;
