+++ striker-ui-api/src/lib/shell/print.ts
@@ -1,0 +1,7 @@
+export const print = (
+  message: string,
+  {
+    eol = '\n',
+    stream = 'stdout',
+  }: { eol?: string; stream?: 'stderr' | 'stdout' } = {},
+) => process[stream].write(`${message}${eol}`);
