+++ striker-ui-api/src/lib/shell/resolveGid.ts
@@ -1,0 +1,3 @@
+import { resolveId } from './resolveId';
+
+export const resolveGid = (id: number | string) => resolveId(id, 'group');
