+++ striker-ui-api/src/lib/shell/resolveId.ts
@@ -1,0 +1,4 @@
+import { getent } from './getent';
+
+export const resolveId = (id: number | string, database: string) =>
+  Number(getent(database, String(id)).split(':', 3)[2]);
