+++ striker-ui-api/src/lib/shell/resolveUid.ts
@@ -1,0 +1,3 @@
+import { resolveId } from './resolveId';
+
+export const resolveUid = (id: number | string) => resolveId(id, 'passwd');
