+++ striker-ui-api/src/lib/shell/uuid.ts
@@ -1,0 +1,3 @@
+import { uuidgen } from './uuidgen';
+
+export const uuid = () => uuidgen('--random').trim();
