+++ striker-ui-api/src/lib/shell/uuidgen.ts
@@ -1,0 +1,6 @@
+import { SERVER_PATHS } from '../consts';
+
+import { systemCall } from './systemCall';
+
+export const uuidgen = (...args: string[]) =>
+  systemCall(SERVER_PATHS.usr.bin.uuidgen.self, args);
